+++ src/ncc/Objects/ProjectConfiguration/Assembly.php
@@ -27,6 +27,9 @@
     use ncc\Enums\RegexPatterns;
     use ncc\Exceptions\ConfigurationException;
     use ncc\Interfaces\BytecodeObjectInterface;
+    use ncc\Interfaces\ValidatableObjectInterface;
+    use ncc\ThirdParty\Symfony\Uid\Uuid;
+    use ncc\ThirdParty\Symfony\Uid\UuidV4;
     use ncc\Utilities\Functions;
     use ncc\Utilities\Validate;
 
@@ -34,7 +37,7 @@
      * @author Zi Xing Narrakas
      * @copyright Copyright (C) 2022-2023. Nosial - All Rights Reserved.
      */
-    class Assembly implements BytecodeObjectInterface
+    class Assembly implements BytecodeObjectInterface, ValidatableObjectInterface
     {
         /**
          * The software name
@@ -96,6 +99,25 @@
          * @var string
          */
         private $uuid;
+
+        /**
+         * Assembly constructor.
+         */
+        public function __construct(string $name, string $package, string $version='1.0.0', ?string $uuid=null)
+        {
+            $this->name = $name;
+            $this->package = $package;
+            $this->version = $version;
+
+            if($uuid === null)
+            {
+                $this->uuid = Uuid::v4()->toRfc4122();
+            }
+            else
+            {
+                $this->uuid = $uuid;
+            }
+        }
 
         /**
          * @return string
@@ -242,105 +264,54 @@
         }
 
         /**
-         * Validates the object information to detect possible errors
-         *
-         * @param bool $throw_exception
-         * @return bool
-         * @throws ConfigurationException
+         * @inheritDoc
          */
-        public function validate(bool $throw_exception=True): bool
+        public function validate(): void
         {
             if(!preg_match(RegexPatterns::UUID, $this->uuid))
             {
-                if($throw_exception)
-                {
-                    throw new ConfigurationException(sprintf('The UUID is not a valid v4 UUID: %s, in property Assembly.UUID', $this->uuid));
-                }
-
-                return false;
+                throw new ConfigurationException(sprintf('The UUID is not a valid v4 UUID: %s, in property assembly.uuid', $this->uuid));
             }
 
             if($this->version !== null && !Validate::version($this->version))
             {
-                if($throw_exception)
-                {
-                    throw new ConfigurationException(sprintf('The version number is invalid: %s, in property Assembly.Version', $this->version));
-                }
-
-                return false;
+                throw new ConfigurationException(sprintf('The version number is invalid: %s, in property assembly.version', $this->version));
             }
 
             if($this->package !== null && !preg_match(RegexPatterns::PACKAGE_NAME_FORMAT, $this->package))
             {
-                if($throw_exception)
-                {
-                    throw new ConfigurationException(sprintf('The package name is invalid: %s, in property Assembly.Package', $this->package));
-                }
-
-                return false;
+                throw new ConfigurationException(sprintf('The package name is invalid: %s, in property assembly.package', $this->package));
             }
 
             if($this->name !== null && strlen($this->name) > 126)
             {
-                if($throw_exception)
-                {
-                    throw new ConfigurationException(sprintf('The name cannot be larger than 126 characters: %s, in property Assembly.Name', $this->name));
-                }
-
-                return false;
+                throw new ConfigurationException(sprintf('The name cannot be larger than 126 characters: %s, in property assembly.name', $this->name));
             }
 
             if($this->description !== null && strlen($this->description) > 512)
             {
-                if($throw_exception)
-                {
-                    throw new ConfigurationException(sprintf('The description cannot be larger than 512 characters: %s, in property Assembly.Description', $this->description));
-                }
-
-                return false;
+                throw new ConfigurationException(sprintf('The description cannot be larger than 512 characters: %s, in property assembly.description', $this->description));
             }
 
             if($this->company !== null && strlen($this->company) > 126)
             {
-                if($throw_exception)
-                {
-                    throw new ConfigurationException(sprintf('The company cannot be larger than 126 characters: %s, in property Assembly.Company', $this->company));
-                }
-
-                return false;
+                throw new ConfigurationException(sprintf('The company cannot be larger than 126 characters: %s, in property assembly.company', $this->company));
             }
 
             if($this->product !== null && strlen($this->product) > 256)
             {
-                if($throw_exception)
-                {
-                    throw new ConfigurationException(sprintf('The product cannot be larger than 256 characters: %s, in property Assembly.Product', $this->product));
-                }
-
-                return false;
+                throw new ConfigurationException(sprintf('The product cannot be larger than 256 characters: %s, in property assembly.product', $this->product));
             }
 
             if($this->copyright !== null && strlen($this->copyright) > 256)
             {
-                if($throw_exception)
-                {
-                    throw new ConfigurationException(sprintf('The copyright cannot be larger than 256 characters: %s, in property Assembly.Copyright', $this->company));
-                }
-
-                return false;
+                throw new ConfigurationException(sprintf('The copyright cannot be larger than 256 characters: %s, in property assembly.copyright', $this->company));
             }
 
             if($this->trademark !== null && strlen($this->trademark) > 256)
             {
-                if($throw_exception)
-                {
-                    throw new ConfigurationException(sprintf('The trademark cannot be larger than 256 characters: %s, in property Assembly.Trademark', $this->trademark));
-                }
-
-                return false;
+                throw new ConfigurationException(sprintf('The trademark cannot be larger than 256 characters: %s, in property assembly.trademark', $this->trademark));
             }
-
-            return true;
         }
 
         /**
@@ -400,20 +371,30 @@
 
         /**
          * @inheritDoc
+         * @throws ConfigurationException
          */
         public static function fromArray(array $data): Assembly
         {
-            $object = new self();
+            $name = Functions::array_bc($data, 'name');
+            $package = Functions::array_bc($data, 'package');
 
-            $object->name = Functions::array_bc($data, 'name');
-            $object->package = Functions::array_bc($data, 'package');
+            if($name === null)
+            {
+                throw new ConfigurationException('The property \'assembly.name\' must not be null.');
+            }
+
+            if($package === null)
+            {
+                throw new ConfigurationException('The property \'assembly.package\' must not be null.');
+            }
+
+            $object = new self($name, $package, Functions::array_bc($data, 'version'), Functions::array_bc($data, 'uuid'));
+
             $object->description = Functions::array_bc($data, 'description');
             $object->company = Functions::array_bc($data, 'company');
             $object->product = Functions::array_bc($data, 'product');
             $object->copyright = Functions::array_bc($data, 'copyright');
             $object->trademark = Functions::array_bc($data, 'trademark');
-            $object->version = Functions::array_bc($data, 'version');
-            $object->uuid = Functions::array_bc($data, 'uuid');
 
             return $object;
         }